Better organize demo files
