--- Demo/Sources/Media.swift
@@ -1,76 +1,0 @@
-//
-//  Copyright (c) SRG SSR. All rights reserved.
-//
-//  License information is available from the LICENSE file.
-//
-
-import Castor
-import Foundation
-import PillarboxPlayer
-
-struct Media: Hashable, Identifiable {
-    enum `Type`: Hashable {
-        case url(URL)
-        case urn(String)
-    }
-
-    let id = UUID()
-    let title: String
-    let imageUrl: URL?
-    let type: Type
-
-    init(title: String, imageUrl: URL? = nil, type: Type) {
-        self.title = title
-        self.imageUrl = imageUrl
-        self.type = type
-    }
-
-    init(from asset: CastAsset) {
-        let title = asset.metadata?.title ?? "Untitled"
-        let imageUrl = asset.metadata?.imageUrl()
-        switch asset.kind {
-        case let .simple(url):
-            self.init(title: title, imageUrl: imageUrl, type: .url(url))
-        case let .custom(urn):
-            self.init(title: title, imageUrl: imageUrl, type: .urn(urn))
-        }
-    }
-
-    func item() -> PlayerItem {
-        switch type {
-        case let .url(url):
-            return .simple(url: url, metadata: self)
-        case let .urn(urn):
-            return .urn(urn)
-        }
-    }
-
-    func asset() -> CastAsset {
-        switch type {
-        case let .url(url):
-            return .simple(url: url, metadata: castMetadata())
-        case let .urn(urn):
-            return .custom(identifier: urn, metadata: castMetadata())
-        }
-    }
-
-    func castMetadata() -> CastMetadata {
-        .init(title: title, image: castImage())
-    }
-
-    private func castImage() -> CastImage? {
-        guard let imageUrl else { return nil }
-        return .init(url: imageUrl)
-    }
-}
-
-extension Media: AssetMetadata {
-    private var imageSource: ImageSource {
-        guard let imageUrl else { return .none }
-        return .url(standardResolution: imageUrl)
-    }
-
-    var playerMetadata: PlayerMetadata {
-        .init(title: title, imageSource: imageSource)
-    }
-}
